ACP-L4-W3-worker-events: keep an older identity retry idempotent after a newer sequence is accepted
PublishRecord rejected an exact retry of an already-accepted
(SourceType, SourceID, SourceSequence, SourceEventID) tuple with
ErrOutOfOrderPublication once a later SourceSequence had advanced the
per-source watermark, instead of resolving it to the original record
as a duplicate the way Events' permanent per-topic dedup requires.
Track accepted identities separately from the ordering watermark so
the regression check only applies to identities never accepted
before. Also splits publish_test.go (pushed over the 1000-line
backendsizecheck limit by the new regression test) into two files by
responsibility.
